+++ examples/echo/main.go
@@ -4,17 +4,17 @@ import (
 	"net/http"
 
 	"github.com/labstack/echo"
-	"github.com/philippgille/ln-paywall/pay"
+	"github.com/philippgille/ln-paywall/wall"
 )
 
 func main() {
 	e := echo.New()
 
 	// Configure and use middleware
-	invoiceOptions := pay.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
-	lndOptions := pay.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
-	storageClient := pay.NewGoMap()
-	e.Use(pay.NewEchoMiddleware(invoiceOptions, lndOptions, storageClient, nil))
+	invoiceOptions := wall.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
+	lndOptions := wall.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
+	storageClient := wall.NewGoMap()
+	e.Use(wall.NewEchoMiddleware(invoiceOptions, lndOptions, storageClient, nil))
 
 	e.GET("/ping", func(c echo.Context) error {
 		return c.String(http.StatusOK, "pong")