+++ examples/gin/main.go
@@ -4,17 +4,17 @@ import (
 	"net/http"
 
 	"github.com/gin-gonic/gin"
-	"github.com/philippgille/ln-paywall/pay"
+	"github.com/philippgille/ln-paywall/wall"
 )
 
 func main() {
 	r := gin.Default()
 
 	// Configure and use middleware
-	invoiceOptions := pay.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
-	lndOptions := pay.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
-	storageClient := pay.NewGoMap()
-	r.Use(pay.NewGinMiddleware(invoiceOptions, lndOptions, storageClient))
+	invoiceOptions := wall.DefaultInvoiceOptions // Price: 1 Satoshi; Memo: "API call"
+	lndOptions := wall.DefaultLNDoptions         // Address: "localhost:10009", CertFile: "tls.cert", MacaroonFile: "invoice.macaroon"
+	storageClient := wall.NewGoMap()
+	r.Use(wall.NewGinMiddleware(invoiceOptions, lndOptions, storageClient))
 
 	r.GET("/ping", func(c *gin.Context) {
 		c.String(http.StatusOK, "pong")