+++ wall/bolt_test.go
@@ -1,0 +1,20 @@
+package wall_test
+
+import (
+	"testing"
+
+	"github.com/philippgille/ln-paywall/wall"
+)
+
+// TestBoltClient tests if the BoltClient struct implements the StorageClient interface.
+// This doesn't happen at runtime, but at compile time.
+func TestBoltClient(t *testing.T) {
+	t.SkipNow()
+	invoiceOptions := wall.InvoiceOptions{}
+	lndOptions := wall.LNDoptions{}
+	boltClient, _ := wall.NewBoltClient(wall.DefaultBoltOptions)
+	wall.NewHandlerFuncMiddleware(invoiceOptions, lndOptions, boltClient)
+	wall.NewHandlerMiddleware(invoiceOptions, lndOptions, boltClient)
+	wall.NewGinMiddleware(invoiceOptions, lndOptions, boltClient)
+	wall.NewEchoMiddleware(invoiceOptions, lndOptions, boltClient, nil)
+}