+++ RELEASES.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 vNext
 -----
 
+### Breaking changes
+
+- Changed: Renamed package from `pay` to `wall`
+
 v0.3.0 (2018-08-12)
 -------------------